07 - Introduction to Keywords and Expressions
Introduction to if-then, else if and else statements.
Also introduced how to place relational expressions in the if-then statements.
Also learned that you cannot access a variable declared within an if-then statement (i.e. 'scope').
You can improve this limitation by declaring the variable first and initializing it with a dummy var / error code.