10.2 - Introduction to Using DiffMerge - Modified
Upload of Modified Java Code to analyze in DiffMerge.
Typically, we load Original Code (stable code) on the left and new modified Code on the right. We can then see differences between the two codes highlighted in red and blue highlighting is also used on the line number.
We can then copy stable code from the left to right, where appropriate.
This is a great tool for checking versions of code. Similarly, git can also show differences between versions with highlighting and changelogs.