Update Main.java - Added Exercise #4
Added Exercise #4: Leap Year Calculator
Uses nested if statements. But, it might be more clear with better logical statements. Leap Year if a year is divisible by 4, 100 and 400 years, but not if divisible by 4 and 100 years only. Also, it is a Leap year if it is divisible by 4 and not 100 or 400 years.