13 - Introduction to Method Overloading
An Introduction to Method Overloading
I had never seen Method Overloading before, but I have known about default variable values and parameters from various modelling / analytics / ML methods from Python. I.e. if you don't specify a particular variable in the Method call, it will be given a default variable value.
In this introductory example, we have a feet and inches to centimeters calculator. The first iteration of the Method, it expects a variable value for Feet and Inches to calculate centimeters. In the second iteration, only an Inches variable is passed, it is converted to Feet and Inches and the first iteration of the Method is called in the second iteration. It was very interesting. I can see how this can be applied to default variable values in Methods.