Update CodingExercises.java 2022-06-19 01
Added Exercise 22 - Perfect Number
The method functions as follows. A while loop steps through 1 to half of the number size. Every step is checked for a proper positive divisor i.e. Number % step == 0.
If == 0, Sum += divisor. Returns true only if the sum of proper positive divisors is equal to the originally input number.