Update CodingExercise.java 2022-06-21 01
Added Exercise 23 - Print Number Digits
This is much more complex than other previously written methods. It uses arrays and a for loop that goes backwards to print out the single digits of a positive number. It also makes use of another method called digitCount to determine the amount of digits that will feed into the starting step of a for loop that counts down from the end to the beginning.