Update Main.java 2022-06-28 01
Added a boolean hasNextInt variable that only proceeds through a following if statement if the user enters an int number. Otherwise, an error message is passed.
Next, the int number is converted to age and checked for validity (0~100, inclusive). If this is true, the user is asked to enter their name (String). Then, their name and age is printed.