15 - Reading User Input Challenge
The user is asked to input an int between 2-15 inclusive. If valid, enters while(true) loop that breaks only when the count == the originally entered int. The while loop will continue to loop only once this number of valid ints has been entered.
There are three error messages, one if a non-int is entered, and one if an int outside of the valid 2-15 range is entered initially, and then if a non-int is entered during the while(true) loop.