16 - Min Max Challenge
Requests User to continuously input int numbers and stores the Min and Max, using a While(true) loop. This breaks when the user inputs a non-int. Min and Max are initialized with zero, and there is a dummy counter variable for the first iteration to initialize the first step with the Min, Max = first int number inputted. From there, simple if statements are used.