17 - Input Calculator
A while (true) loop asks for int numbers to be continuously input. Breaks/exits on non-int being entered.
On break, the class calculates sum and average of numbers.
Requires some conversion of double to int for average to work.
Sum and Average are initialized with 0, but calculate normally as long as at least 1 int number has been input.