14 - Static vs Instance Methods
14 - Static vs Instance Methods
Discusses and explores Static vs Instance methods and when to use them. In general, it appears that static methods can only be used when there are not class specific variables. I.e. A basic calculation or input variables to a method should be a Static method, but a class with unique values and parameters should be an Instance method.