2023-06-03 - Queue Example Class
An example of creating a Queue class and objects which is an array of n storage with extra properties such as queue length. Queue object properties are private and can only be modified with public methods. Methods include: a constructor, removing the first element of the queue (passing) O(n) due to value shifting, adding to the end of the queue O(1), and printing elements of the queue object.