23-07-04 Update 0A_intro.ipynb
Added section on Reshaping, views, stacking, squeezing and unsqueezing tensors. Demonstrates how these methods and functions work on tensors.
Reshape and view are similar, but view strictly points to the original tensor, so causes mutation. Reshape, if possible, will create a view and cause mutation. If it is not possible to create a view, reshape will instead create a copy of the original tensor. This can be circumvented with torch.clone(target_tensor).reshape((target_dims)).
Stacking with dim, vstack and stack, stack tensors together.
torch.squeeze and torch.unsqueeze are particularly useful at deleting and adding 1-dimensions to the target tensor, but offer further functionality with the dim parameter.